refactor: ノイズコンポーネントを ui からルート内のコンポーネントに移動させた
