+++ src/app/_components/Noise/Noise.tsx
