+++ src/app/_components/Noise/noise.css.ts
